finopti-platform/sub_agents/storage_agent_adk/requirements.txt
flask>=3.0.3
gunicorn>=22.0.0
google-cloud-aiplatform>=1.38.0
google-cloud-secret-manager>=2.19.0
google-cloud-resource-manager>=1.12.3
google-adk>=1.21.0
google-genai>=0.6.0
google-generativeai>=0.8.0
requests>=2.32.4
python-dotenv>=1.0.1
structlog>=24.1.0
Flask>=3.0.0
google-cloud-storage>=2.14.0
# Observability
arize-phoenix>=4.0.0
openinference-instrumentation-google-adk>=0.1.9
openinference-semantic-conventions>=0.1.9
opentelemetry-sdk
opentelemetry-exporter-otlp
